+++ service/internal/bootstrap/bootstrap.go
@@ -72,6 +72,11 @@ func (b *Bootstrap) Run(ctx context.Context) error {
 		return microerror.Mask(err)
 	}
 
+	err = b.createReadDefaultCatalogsRole(ctx)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
 	err = b.createWriteAllClusterRoleBindingToGSGroup(ctx)
 	if err != nil {
 		return microerror.Mask(err)