+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+
+- Add creation of `read-default-catalogs` Role.
 - Add creation of `read-releases` ClusterRole.
 
 ## [0.21.0] - 2022-02-16